+++ spec/lib/payrix/client_spec.rb
@@ -37,7 +37,7 @@ RSpec.describe Payrix::Client do
           WebMock
             .stub_request(
               :get,
-              'https://test-api.payrix.com/txns?page[number]=1&page[limit]=10&expand[merchant][]=',
+              'https://api.payrix.com/txns?page[number]=1&page[limit]=10&expand[merchant][]=',
             )
             .with(
               headers: {
@@ -68,6 +68,7 @@ RSpec.describe Payrix::Client do
                 limit: 10,
                 expand: ['merchant'],
                 api_key: 'custom-key',
+                environment: :production,
               },
             )
 
@@ -110,7 +111,7 @@ RSpec.describe Payrix::Client do
       it 'supports advanced requests' do
         stub =
           WebMock
-            .stub_request(:post, 'https://test-api.payrix.com/txns?expand[merchant][]=')
+            .stub_request(:post, 'https://api.payrix.com/txns?expand[merchant][]=')
             .with(
               headers: {
                 'Content-Type' => 'application/json',
@@ -142,6 +143,7 @@ RSpec.describe Payrix::Client do
               options: {
                 expand: ['merchant'],
                 api_key: 'custom-key',
+                environment: :production,
               },
             )
 
@@ -192,7 +194,7 @@ RSpec.describe Payrix::Client do
       it 'supports advanced requests' do
         stub =
           WebMock
-            .stub_request(:put, 'https://test-api.payrix.com/txns/t1_txn_64026b07cc6a79dd5cfd0da')
+            .stub_request(:put, 'https://api.payrix.com/txns/t1_txn_64026b07cc6a79dd5cfd0da')
             .with(
               headers: {
                 'Content-Type' => 'application/json',
@@ -223,6 +225,7 @@ RSpec.describe Payrix::Client do
               },
               options: {
                 api_key: 'custom-key',
+                environment: :production,
               },
             )
 
@@ -261,7 +264,7 @@ RSpec.describe Payrix::Client do
       it 'supports advanced requests' do
         stub =
           WebMock
-            .stub_request(:delete, 'https://test-api.payrix.com/txns/t1_txn_64026b07cc6a79dd5cfd0da')
+            .stub_request(:delete, 'https://api.payrix.com/txns/t1_txn_64026b07cc6a79dd5cfd0da')
             .with(
               headers: {
                 'Content-Type' => 'application/json',
@@ -286,6 +289,7 @@ RSpec.describe Payrix::Client do
               resource: 'txns/t1_txn_64026b07cc6a79dd5cfd0da',
               options: {
                 api_key: 'custom-key',
+                environment: :production,
               },
             )